Update README.md
Project Report: Evaluation of Artificial Neural Network for Credit Risk Prediction

Introduction:

This report presents the results and evaluation of a data science project conducted with an artificial neural network for bank credit risk prediction. The project involved data analysis, model development, and performance evaluation.


Data Preparation and Preliminary Assessment:

Analyses conducted on the dataset revealed that over 80% of duplicate data was present and there was insufficient variation among variables. This was identified as a significant limitation for building a successful model. Following necessary data preparation and preliminary assessments, it was concluded that the dataset was not suitable for constructing a successful model.


Model Evaluation:

The trained artificial neural network model was evaluated on the test dataset. The obtained results are as follows:

Test Accuracy: 45.7%

Test Precision: 45.6%

Test Recall: 99.5%

Test F1: 0.625%


Results and Discussion:

According to the obtained results, it can be seen that the model's recall value is quite high, but accuracy, precision, and loss values are low. This indicates that the model is quite successful in identifying negative situations in credit risk prediction but weak in identifying positive situations.


Recommendations:

The high recall value of the model indicates its success in identifying negative situations. However, for better performance in identifying positive situations, the model needs to be optimized. More observations should be added to the dataset, and more explanatory variables should be used. The hyperparameters and architecture of the model should be adjusted in more detail.


Conclusion:

The results of this project demonstrate the model's ability to identify both positive and negative situations. However, further development of the model and expansion of the dataset are necessary to achieve better performance.

The test recall (%99.5) indicates how accurately the model predicts true positives. In other words, this value means that the model correctly identified or classified 99.5% of the true positives. This demonstrates that the model has a very high success rate in predicting credit risk, accurately identifying the true credit-risk customers.